fixed the cards to only render once
